feat: add playerService.ts
